+++ src/main/java/utcapitole/miage/bloop/service/UtilisateurService.java
@@ -2,6 +2,7 @@ package utcapitole.miage.bloop.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import utcapitole.miage.bloop.model.entity.Utilisateur;
 import utcapitole.miage.bloop.repository.UtilisateurRepository;
