+++ src/test/java/utcapitole/miage/bloop/controller/ProfilControllerTest.java
@@ -1,0 +1,51 @@
+package utcapitole.miage.bloop.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+import utcapitole.miage.bloop.model.entity.Utilisateur;
+import utcapitole.miage.bloop.service.UtilisateurService;
+
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+@WebMvcTest(ProfilController.class)
+@ActiveProfiles("test")
+@AutoConfigureMockMvc(addFilters = false)
+class ProfilControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockitoBean
+    private UtilisateurService utilisateurService;
+
+    @Test
+    void testAfficherMonProfil_Succes() throws Exception {
+        Utilisateur utilisateur = new Utilisateur();
+        utilisateur.setIdUser(1L);
+        utilisateur.setNomUser("Test");
+
+        when(utilisateurService.getUtilisateurConnecte()).thenReturn(utilisateur);
+        when(utilisateurService.getUtilisateurParId(1L)).thenReturn(utilisateur);
+
+        mockMvc.perform(get("/profil/voirProfil"))
+                .andExpect(status().isOk())
+                .andExpect(model().attributeExists("utilisateur"))
+                .andExpect(view().name("voirProfil"));
+    }
+
+    @Test
+    void testAfficherMonProfil_RedirectionAccueil() throws Exception {
+        when(utilisateurService.getUtilisateurConnecte()).thenReturn(null);
+
+        mockMvc.perform(get("/profil/voirProfil"))
+                .andExpect(status().isOk())
+                .andExpect(view().name("accueil"));
+    }
+}
